decimals fixed!

+++ Assets/Objects/ActiveObjects/Cat.cs
@@ -33,7 +33,7 @@ namespace Objects.ActiveObjects
 
         protected override void ConfigureAndPlayMilked(Transform transformMe)
         {
-            Decimal moneyMoney = objectCounter * productionPower * (MoneyManagerSingleton.instance.numberOfTitties + 1);
+            Decimal moneyMoney = (Decimal)objectCounter * productionPower * (Decimal)(MoneyManagerSingleton.instance.numberOfTitties + 1);
             StartCoroutine(PlayMilkedCoroutine(transformMe,
                 Helpers.GetObjectPositionRelativeToCanvas(transformMe.position), moneyMoney));
         }
